Extract S3LogDir to it's own module.
